requirements.txt
PyQt5
psutil
pywin32
requests
Pillow
pystray